Can connect to non PPN servers
I added an abstraction for reading and writing to sockets on the
respective threads. This allows for a PPN client to connect to a Non-PPN
server and vice versa. (This only works if the NON PPN client/server
send message size then message in bytes). Also all messages sent are
bytes